Packages e classes refatoradas para melhor entendimento
