+++ src/main/java/com/example/apicasadocodigo/categoria/Categoria.java
@@ -1,0 +1,27 @@
+package com.example.apicasadocodigo.categoria;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.validation.constraints.NotBlank;
+
+@Entity
+public class Categoria {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    private @NotBlank String nome;
+
+    @Deprecated
+    public Categoria() {
+    }
+
+    public Categoria(@NotBlank String nome) {
+        this.nome = nome;
+    }
+
+    public String getNome() {
+        return nome;
+    }
+}
