--- src/main/java/com/example/apicasadocodigo/consultalivro/ConsultaLivroController.java
@@ -1,26 +1,0 @@
-package com.example.apicasadocodigo.consultalivro;
-
-import com.example.apicasadocodigo.novolivro.Livro;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.util.List;
-
-@RestController
-public class ConsultaLivroController {
-    @Autowired
-    private LivroRepository livroRepository;
-
-    @GetMapping("/livros")
-    public ResponseEntity<List<Livro>> listaLivros() {
-        return ResponseEntity.ok().body(livroRepository.findAll());
-    }
-
-    @GetMapping("/livros/{id}")
-    public ResponseEntity<Livro> buscaLivroPorId(@PathVariable Long id) {
-        return ResponseEntity.ok().body(livroRepository.findById(id).get());
-    }
-}
