--- src/main/java/com/example/apicasadocodigo/cupom/Cupom.java
@@ -1,68 +1,0 @@
-package com.example.apicasadocodigo.cupom;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.validation.constraints.Future;
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
-import javax.validation.constraints.Positive;
-import java.math.BigDecimal;
-import java.time.LocalDate;
-
-@Entity
-public class Cupom {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    private @NotBlank String codigo;
-    private @Positive @NotNull BigDecimal desconto;
-    private @Future @NotNull LocalDate validade;
-
-    @Deprecated
-    public Cupom() {
-    }
-
-    public Cupom(@NotNull String codigo, @Positive @NotNull BigDecimal desconto, @Future @NotNull LocalDate validade) {
-        this.codigo = codigo;
-        this.desconto = desconto;
-        this.validade = validade;
-    }
-
-    public boolean valido() {
-        return LocalDate.now().compareTo(this.validade) <= 0;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getCodigo() {
-        return codigo;
-    }
-
-    public void setCodigo(String codigo) {
-        this.codigo = codigo;
-    }
-
-    public BigDecimal getDesconto() {
-        return desconto;
-    }
-
-    public void setDesconto(BigDecimal desconto) {
-        this.desconto = desconto;
-    }
-
-    public LocalDate getValidade() {
-        return validade;
-    }
-
-    public void setValidade(LocalDate validade) {
-        this.validade = validade;
-    }
-}
