--- src/main/java/com/example/apicasadocodigo/cupom/CupomRepository.java
@@ -1,9 +1,0 @@
-package com.example.apicasadocodigo.cupom;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface CupomRepository extends JpaRepository<Cupom, Long> {
-    public Cupom findByCodigo(String codigo);
-}
