+++ src/main/java/com/example/apicasadocodigo/fluxocompra/compra/CompraResponse.java
@@ -1,0 +1,162 @@
+package com.example.apicasadocodigo.fluxocompra.compra;
+
+import com.example.apicasadocodigo.fluxocompra.itemcarrinho.ItemCarrinho;
+
+import java.math.BigDecimal;
+import java.util.Set;
+
+public class CompraResponse {
+    private String email;
+    private String nome;
+    private String sobrenome;
+    private String documento;
+    private String endereco;
+    private String complemento;
+    private String cidade;
+    private String pais;
+    private String estado;
+    private String telefone;
+    private String cep;
+    private Set<ItemCarrinho> itens;
+    private BigDecimal total;
+    private BigDecimal desconto;
+    private BigDecimal totalComDesconto;
+
+    public CompraResponse(Compra compra) {
+        this.email = compra.getEmail();
+        this.nome = compra.getNome();
+        this.sobrenome = compra.getSobrenome();
+        this.documento = compra.getDocumento();
+        this.endereco = compra.getEndereco();
+        this.complemento = compra.getComplemento();
+        this.cidade = compra.getCidade();
+        this.pais = compra.getPais().getNome();
+        this.estado = compra.getEstado().getNome();
+        this.telefone = compra.getTelefone();
+        this.cep = compra.getCep();
+        this.itens = compra.getPedido().getItens();
+        this.total = compra.total();
+        this.desconto = compra.getCupom().getDesconto();
+        this.totalComDesconto = compra.calcularTotalComDesconto();
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    public String getNome() {
+        return nome;
+    }
+
+    public void setNome(String nome) {
+        this.nome = nome;
+    }
+
+    public String getSobrenome() {
+        return sobrenome;
+    }
+
+    public void setSobrenome(String sobrenome) {
+        this.sobrenome = sobrenome;
+    }
+
+    public String getDocumento() {
+        return documento;
+    }
+
+    public void setDocumento(String documento) {
+        this.documento = documento;
+    }
+
+    public String getEndereco() {
+        return endereco;
+    }
+
+    public void setEndereco(String endereco) {
+        this.endereco = endereco;
+    }
+
+    public String getComplemento() {
+        return complemento;
+    }
+
+    public void setComplemento(String complemento) {
+        this.complemento = complemento;
+    }
+
+    public String getCidade() {
+        return cidade;
+    }
+
+    public void setCidade(String cidade) {
+        this.cidade = cidade;
+    }
+
+    public String getPais() {
+        return pais;
+    }
+
+    public void setPais(String pais) {
+        this.pais = pais;
+    }
+
+    public String getEstado() {
+        return estado;
+    }
+
+    public void setEstado(String estado) {
+        this.estado = estado;
+    }
+
+    public String getTelefone() {
+        return telefone;
+    }
+
+    public void setTelefone(String telefone) {
+        this.telefone = telefone;
+    }
+
+    public String getCep() {
+        return cep;
+    }
+
+    public void setCep(String cep) {
+        this.cep = cep;
+    }
+
+    public Set<ItemCarrinho> getItens() {
+        return itens;
+    }
+
+    public void setItens(Set<ItemCarrinho> itens) {
+        this.itens = itens;
+    }
+
+    public BigDecimal getTotal() {
+        return total;
+    }
+
+    public void setTotal(BigDecimal total) {
+        this.total = total;
+    }
+
+    public BigDecimal getDesconto() {
+        return desconto;
+    }
+
+    public void setDesconto(BigDecimal desconto) {
+        this.desconto = desconto;
+    }
+
+    public BigDecimal getTotalComDesconto() {
+        return totalComDesconto;
+    }
+
+    public void setTotalComDesconto(BigDecimal totalComDesconto) {
+        this.totalComDesconto = totalComDesconto;
+    }
+}
