+++ src/main/java/com/example/apicasadocodigo/fluxocompra/itemcarrinho/ItemCarrinhoResponse.java
@@ -1,0 +1,39 @@
+//package com.example.apicasadocodigo.fluxocompra.itemcarrinho;
+//
+//import java.math.BigDecimal;
+//
+//public class ItemCarrinhoResponse {
+//    private String livro;
+//    private int quantidade;
+//    private BigDecimal preco;
+//
+//    public ItemCarrinhoResponse(ItemCarrinho itemCarrinho) {
+//        this.livro = itemCarrinho.getLivro().getTitulo();
+//        this.quantidade = itemCarrinho.getQuantidade();
+//        this.preco = itemCarrinho.getPreco();
+//    }
+//
+//    public String getLivro() {
+//        return livro;
+//    }
+//
+//    public void setLivro(String livro) {
+//        this.livro = livro;
+//    }
+//
+//    public int getQuantidade() {
+//        return quantidade;
+//    }
+//
+//    public void setQuantidade(int quantidade) {
+//        this.quantidade = quantidade;
+//    }
+//
+//    public BigDecimal getPreco() {
+//        return preco;
+//    }
+//
+//    public void setPreco(BigDecimal preco) {
+//        this.preco = preco;
+//    }
+//}
