+++ src/main/java/com/example/apicasadocodigo/livro/Livro.java
@@ -1,0 +1,78 @@
+package com.example.apicasadocodigo.livro;
+
+import com.example.apicasadocodigo.categoria.Categoria;
+import com.example.apicasadocodigo.autor.Autor;
+
+import javax.persistence.*;
+import javax.validation.Valid;
+import javax.validation.constraints.*;
+import java.math.BigDecimal;
+import java.time.LocalDate;
+
+@Entity
+public class Livro {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    private @NotBlank String titulo;
+    private @NotBlank @Size(max = 500) String resumo;
+    private @NotBlank String sumario;
+    private @NotNull @Min(20) BigDecimal preco;
+    private @Min(100) int numeroDePaginas;
+    private @NotBlank String isbn;
+    private @NotNull @Future LocalDate dataDePublicacao;
+    private @ManyToOne @NotNull @Valid Autor autor;
+    private @ManyToOne @NotNull @Valid Categoria categoria;
+
+    @Deprecated
+    public Livro() {
+    }
+
+    public Livro(@NotBlank String titulo,
+                 @NotBlank @Size(max = 500) String resumo, @NotBlank String sumario,
+                 @NotNull @Min(20) BigDecimal preco, @Min(100) int numeroDePaginas,
+                 @NotBlank String isbn, @NotNull @Future LocalDate dataDePublicacao,
+                 @NotNull @Valid Autor autor, @NotNull @Valid Categoria categoria) {
+        this.titulo = titulo;
+        this.resumo = resumo;
+        this.sumario = sumario;
+        this.preco = preco;
+        this.numeroDePaginas = numeroDePaginas;
+        this.isbn = isbn;
+        this.dataDePublicacao = dataDePublicacao;
+        this.autor = autor;
+        this.categoria = categoria;
+    }
+
+    public String getTitulo() {
+        return titulo;
+    }
+
+    public String getResumo() {
+        return resumo;
+    }
+
+    public BigDecimal getPreco() {
+        return preco;
+    }
+
+    public int getNumeroDePaginas() {
+        return numeroDePaginas;
+    }
+
+    public String getIsbn() {
+        return isbn;
+    }
+
+    public LocalDate getDataDePublicacao() {
+        return dataDePublicacao;
+    }
+
+    public Autor getAutor() {
+        return autor;
+    }
+
+    public Categoria getCategoria() {
+        return categoria;
+    }
+}
