+++ src/main/java/com/example/apicasadocodigo/livro/LivroRepository.java
@@ -1,0 +1,9 @@
+package com.example.apicasadocodigo.livro;
+
+import com.example.apicasadocodigo.livro.Livro;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface LivroRepository extends JpaRepository<Livro, Long> {
+}
