+++ src/main/java/com/example/apicasadocodigo/livro/NovoLivroRequest.java
@@ -1,0 +1,116 @@
+package com.example.apicasadocodigo.livro;
+
+import com.example.apicasadocodigo.compartilhado.UniqueValue;
+import com.example.apicasadocodigo.categoria.Categoria;
+import com.example.apicasadocodigo.autor.Autor;
+import com.fasterxml.jackson.annotation.JsonFormat;
+
+import javax.persistence.EntityManager;
+import javax.validation.constraints.*;
+import java.math.BigDecimal;
+import java.time.LocalDate;
+
+public class NovoLivroRequest {
+    @NotBlank
+    @UniqueValue(domainClass = Livro.class, fieldName = "titulo")
+    private String titulo;
+    @NotBlank
+    @Size(max = 500)
+    private String resumo;
+    @NotBlank
+    private String sumario;
+    @NotNull
+    @Min(20)
+    private BigDecimal preco;
+    @Min(100)
+    private int numeroDePaginas;
+    @NotBlank
+    @UniqueValue(domainClass = Livro.class, fieldName = "isbn", message = "Este identificador já existe.")
+    private String isbn;
+    @Future
+    @NotNull
+    @JsonFormat(pattern = "dd/MM/yyyy", shape = JsonFormat.Shape.STRING)
+    private LocalDate dataDePublicacao;
+    @NotNull
+    private Long idAutor;
+    @NotNull
+    private Long idCategoria;
+
+    public Livro toModel(EntityManager manager) {
+        @NotNull Autor autor = manager.find(Autor.class, idAutor);
+        @NotNull Categoria categoria = manager.find(Categoria.class, idCategoria);
+        return new Livro(titulo, resumo, sumario, preco, numeroDePaginas, isbn, dataDePublicacao, autor, categoria);
+    }
+
+    public String getTitulo() {
+        return titulo;
+    }
+
+    public void setTitulo(String titulo) {
+        this.titulo = titulo;
+    }
+
+    public String getResumo() {
+        return resumo;
+    }
+
+    public void setResumo(String resumo) {
+        this.resumo = resumo;
+    }
+
+    public String getSumario() {
+        return sumario;
+    }
+
+    public void setSumario(String sumario) {
+        this.sumario = sumario;
+    }
+
+    public BigDecimal getPreco() {
+        return preco;
+    }
+
+    public void setPreco(BigDecimal preco) {
+        this.preco = preco;
+    }
+
+    public int getNumeroDePaginas() {
+        return numeroDePaginas;
+    }
+
+    public void setNumeroDePaginas(int numeroDePaginas) {
+        this.numeroDePaginas = numeroDePaginas;
+    }
+
+    public String getIsbn() {
+        return isbn;
+    }
+
+    public void setIsbn(String isbn) {
+        this.isbn = isbn;
+    }
+
+    public LocalDate getDataDePublicacao() {
+        return dataDePublicacao;
+    }
+
+    public void setDataDePublicacao(LocalDate dataDePublicacao) {
+        this.dataDePublicacao = dataDePublicacao;
+    }
+
+    public Long getIdAutor() {
+        return idAutor;
+    }
+
+    public void setIdAutor(Long idAutor) {
+        this.idAutor = idAutor;
+    }
+
+    public Long getIdCategoria() {
+        return idCategoria;
+    }
+
+    public void setIdCategoria(Long idCategoria) {
+        this.idCategoria = idCategoria;
+    }
+}
