--- src/main/java/com/example/apicasadocodigo/novoautor/NovoAutorRequest.java
@@ -1,31 +1,0 @@
-package com.example.apicasadocodigo.novoautor;
-
-import com.example.apicasadocodigo.compartilhado.UniqueValue;
-
-import javax.validation.constraints.Email;
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.Size;
-
-public class NovoAutorRequest {
-    @NotBlank
-    private String nome;
-    @NotBlank
-    @Email
-    @UniqueValue(domainClass = Autor.class, fieldName = "email", message = "Este email já está cadastrado.")
-    private String email;
-    @NotBlank
-    @Size(max = 400)
-    private String descricao;
-
-    public NovoAutorRequest(@NotBlank String nome, @NotBlank @Email String email,
-                            @NotBlank @Size(max = 400) String descricao) {
-        super();
-        this.nome = nome;
-        this.email = email;
-        this.descricao = descricao;
-    }
-
-    public Autor toModel() {
-        return new Autor(this.nome, this.email, this.descricao);
-    }
-}
